backend/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
python-multipart
pyjwt
passlib
bcrypt==3.2.0
python-dotenv
email-validator
gunicorn